KI_testing/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-jinja2==1.6
aiosignal==1.4.0
attrs==25.4.0
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.2
frozenlist==1.8.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
linkify-it-py==2.0.3
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdit-py-plugins==0.5.0
mdurl==0.1.2
msgpack==1.1.2
multidict==6.7.1
pillow==12.1.1
platformdirs==4.5.1
propcache==0.4.1
Pygments==2.19.2
PyPDF2==3.0.1
python-docx==1.2.0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
rich==14.3.2
textual==7.5.0
textual-dev==1.8.0
textual-serve==1.1.3
typing_extensions==4.15.0
uc-micro-py==1.0.3
urllib3==2.6.3
Werkzeug==3.1.5
yarl==1.22.0
